+++ src/Components/MainMenuStack/StartStack/MainGameStack/shared/ComponentBottom/index.js
@@ -1,0 +1,1 @@
+export { default as ComponentBottom } from './ComponentBottom';